use float value in response object

+++ src/currency/currencyService.py
@@ -17,10 +17,10 @@ class CurrencyService:
         self.cache_filename = cache_filename
         self.cache_ttl = cache_ttl
 
-    async def convert(self, amount: int, init_currency: Currency, target_currency: Currency) -> float:
+    async def convert(self, amount: float, init_currency: Currency, target_currency: Currency) -> float:
 
             excp_msg = None
-            if not isinstance(amount, int):
+            if not isinstance(amount, float):
                 excp_msg = 'The amount must be a numeric value!'
             elif amount <= 0:
                 excp_msg = 'The amount must be greater than zero'
@@ -33,12 +33,12 @@ class CurrencyService:
             CACHE_KEY = str(init_currency) + str(target_currency) + str(amount)
             
             result = simple_cache.load_key(self.cache_filename, CACHE_KEY)
+            
             if result:
                 log.info('Found in cache !' + target_currency.value + '  rates for ' + init_currency.value + ' ' + str(amount) )
                 return result
 
             try:
-                # If the response was successful, no Exception will be raised
                 params = {'to': target_currency.value,'from': init_currency.value, 'amount': amount}
                 headers = {'apikey': self.api_key}
                 response = requests.get(self.api_endpoint, params, headers=headers)
